feat(design-system): add VisuallyHidden component
