+++ packages/design-system/src/components/VisuallyHidden/index.ts
@@ -1,0 +1,2 @@
+export { VisuallyHidden, useVisuallyHiddenStyles } from './VisuallyHidden';
+export type { VisuallyHiddenProps } from './VisuallyHidden';
